BASE_URL tarkistus lisätty backend kutsuille
